fixing up compute project
